pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "analauracpaiva", email = "analauracpaiva26@gmail.com"},
]
dependencies = [
    "django>=4.2.3",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.2.0",
    "setuptools>=68.1.2",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
    "django-extensions>=3.2.3",
    "pydotplus>=2.0.2",
    "drf-spectacular>=0.26.4",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
]
